Fix race condition and android bugs with incomplete/corrupted installation
This will make sure the www folder is never left in a corrupted state. If it is corrupted, it will always be recreated on app start.

Also fixes a race condition with two installations being started concurrently, and interfering with each other's files. For example both would compete over the www_backup folder, and in my app's case, this would always fail the update because I have a lot of files in the app and file system errors would occur.